Set C/C++ standards through target properties
I replaced target_compile_feature with target properties.

When requiring a language standard, target_compile_feature only ensure a
compiler mode at minimum aware of that standard, but could be greater.
As a result, it only adds '-std=' flag when necessary. This does not
enforce a language standard on the code, and may cause confusion for
CMake novices. Furthermore, the lack of '-std=' flag also breaks some
IDE parsers relying on it (ex. KDevelop).